Use size_t for indexing and sizes

+++ src/darr.c
@@ -8,13 +8,13 @@ void darr_global_set_realloc(darr_realloc_t f);
 
 void darr_global_set_free(darr_free_t f);
 
-size_t darr_data_index(struct darr *d, int i);
+size_t darr_data_index(struct darr *d, size_t i);
 
 void darr_init(struct darr *d, size_t item_size);
 
 void darr_deinit(struct darr *d);
 
-int darr_size(struct darr *d);
+size_t darr_size(struct darr *d);
 
 int darr_resize(struct darr *d, size_t size);
